# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Pametne Kuće API",
        "description": "API dokumentacija za upravljanje uređajima, korisnicima i ulogama (Admin, Običan, Dete).",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalni razvojni server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Prijava korisnika",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uspešna prijava",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Pogrešni podaci"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Odjava korisnika",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Uspešna odjava"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Registracija novog korisnika",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ime": {
                                        "type": "string",
                                        "example": "Marko"
                                    },
                                    "prezime": {
                                        "type": "string",
                                        "example": "Markovic"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "marko123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "uloga": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "dete",
                                            "obican"
                                        ],
                                        "example": "obican"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Korisnik uspešno kreiran"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            }
        },
        "/api/korisnici": {
            "get": {
                "tags": [
                    "Korisnici"
                ],
                "summary": "Prikaz svih korisnika",
                "operationId": "5061f6d16eb244e5b6757a44af053383",
                "responses": {
                    "200": {
                        "description": "Uspešno učitana lista korisnika"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Korisnici"
                ],
                "summary": "Kreiranje novog korisnika",
                "operationId": "b09c5a6427c0ae9a7ce73389ef77acf3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ime": {
                                        "type": "string",
                                        "example": "Petar"
                                    },
                                    "prezime": {
                                        "type": "string",
                                        "example": "Petrović"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "petar123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "lozinka123"
                                    },
                                    "uloga": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "obican",
                                            "dete"
                                        ],
                                        "example": "obican"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Korisnik uspešno kreiran"
                    },
                    "422": {
                        "description": "Greška pri validaciji"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/korisnici/{id}": {
            "get": {
                "tags": [
                    "Korisnici"
                ],
                "summary": "Prikaz detalja jednog korisnika",
                "operationId": "10e4efcd04abd49ded36c98531c4ea4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID korisnika",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno pronađen korisnik"
                    },
                    "404": {
                        "description": "Korisnik nije pronađen"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Korisnici"
                ],
                "summary": "Izmena podataka o korisniku",
                "operationId": "ce20b76969d640b2ad7066574cc652e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID korisnika",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ime": {
                                        "type": "string"
                                    },
                                    "prezime": {
                                        "type": "string"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "uloga": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "obican",
                                            "dete"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Podaci uspešno ažurirani"
                    },
                    "404": {
                        "description": "Korisnik nije pronađen"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Korisnici"
                ],
                "summary": "Brisanje korisnika",
                "operationId": "fb63425b86922025564614af57838826",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID korisnika",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Korisnik uspešno obrisan"
                    },
                    "404": {
                        "description": "Korisnik nije pronađen"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Ovde unesite access_token dobijen nakon logovanja.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}